Safari-611.1.10/LayoutTests/editing/text-iterator/count-matches-in-frames-expected.txt

findme0

findme1
findme2

findme3

PASS Search from frame in normal tree
PASS Search from frame in display:none subtree
PASS Search from frame in zero sized subtree
PASS Search from frame in zero sized subtree with overflow:hidden
